for test//

+++ minuku2-extended/src/main/java/edu/umich/si/inteco/minuku/streamgenerator/AnnotatedImageStreamGenerator.java
@@ -112,7 +112,7 @@ public class AnnotatedImageStreamGenerator<T extends AnnotatedImageDataRecord>
     }
 
     @Override
-    public double getUpdateFrequency() {
+    public long getUpdateFrequency() {
         return -1;
     }
