+++ minuku2-extended/src/main/java/edu/umich/si/inteco/minuku/streamgenerator/ImageStreamGenerator.java
@@ -89,7 +89,7 @@ public class ImageStreamGenerator extends AndroidStreamGenerator<ImageDataRecord
     }
 
     @Override
-    public double getUpdateFrequency() {
+    public long getUpdateFrequency() {
         return -1;
     }
