+++ minuku2-extended/src/main/java/edu/umich/si/inteco/minuku/streamgenerator/LocationStreamGenerator.java
@@ -178,8 +178,8 @@ public class LocationStreamGenerator extends AndroidStreamGenerator<LocationData
     }
 
     @Override
-    public double getUpdateFrequency() {
-        return 15.00; // 1 minutes
+    public long getUpdateFrequency() {
+        return 15; // 1 minutes
     }
 
     @Override
